refactor: restructured files into separate modules
