+++ app/database/__init__.py
@@ -1,0 +1,1 @@
+from app.database.connection import engine, SessionLocal
