+++ app/models/__init__.py
@@ -1,0 +1,1 @@
+from app.models.trips import Trips, TripDays, Activities, Base
