+++ app/models/trips.py
